sunrise sunset data and view, add table region for places in future

+++ app/src/main/java/local/kilg/fw/network/rest/pojo/SnowNight.kt
@@ -7,8 +7,8 @@ import com.google.gson.annotations.SerializedName
 data class SnowNight(
 
 	@field:SerializedName("in")
-	val val_in: Int? = null,
+	val val_in: Double? = null,
 
 	@field:SerializedName("cm")
-	val cm: Int? = null
+	val cm: Double? = null
 )
